train.py: remove visualizer
This script assumes that we run a visdom server, else it insists on
starting one by itself, which is quite annoying.

Also, the visualizer instance is responsible for writing the loss log
file. We write that file independently of visualizer. Additionally, we
remove it before each run instead of appending to it.